RAG/requirements.txt
flask
numpy
python-dotenv
requests
scikit-learn
pypdf
python-docx
sentence-transformers
